+++ libsq/core/oncecache/example/staffdir/cache.go
@@ -2,7 +2,7 @@ package staffdir
 
 import (
 	"context"
-	"github.com/neilotoole/sq/libsq/core/ocache"
+	"github.com/neilotoole/sq/libsq/core/oncecache"
 	"log/slog"
 )
 
@@ -13,11 +13,11 @@ func NewDirCache(log *slog.Logger, db DB) *DirCache {
 		log:   log,
 		db:    db,
 		stats: NewStats(),
-		companies: ocache.New[string, *Company](func(ctx context.Context, _ string) (val *Company, err error) {
+		companies: oncecache.New[string, *Company](func(ctx context.Context, _ string) (val *Company, err error) {
 			return db.GetCompany(ctx)
 		}),
-		depts:     ocache.New[string, *Department](db.GetDepartment),
-		employees: ocache.New[int, *Employee](db.GetEmployee),
+		depts:     oncecache.New[string, *Department](db.GetDepartment),
+		employees: oncecache.New[int, *Employee](db.GetEmployee),
 	}
 }
 
@@ -25,9 +25,9 @@ type DirCache struct {
 	log       *slog.Logger
 	db        DB
 	stats     *Stats
-	companies *ocache.Cache[string, *Company]
-	depts     *ocache.Cache[string, *Department]
-	employees *ocache.Cache[int, *Employee]
+	companies *oncecache.Cache[string, *Company]
+	depts     *oncecache.Cache[string, *Department]
+	employees *oncecache.Cache[int, *Employee]
 }
 
 func (dc *DirCache) Stats() *Stats {
