+++ libsq/core/oncecache/example/staffdir/testdata/acme.json
