+++ libsq/core/oncecache/opt.go
@@ -1,4 +1,4 @@
-package ocache
+package oncecache
 
 import "context"
